Bug in simulation duration when using LAB units.
The summary 'length' functions report the results in the time unit used in
simulation output. This is mostly 'days', but in the case of lab units it is
'hours'. The ecl_sum_data_get_simulation_length( ) would previously return days
anyways, this is now fixed and a test has been updated.

Observe that in function signatures variables are named 'days', where more
correct variable name would be 'duration-in-whatever-time-unit-is-in-action'.