ecl_file_open() - will not fail on broken files.
In the case of a broken file the ecl_file_open() function will continue indexing
the file until a broken ecl_kw instance is found on disk. It will then return a
ecl_file instance which ignores the trailing garbage.

This implies that calling scope, e.g. the ecl_grid or ecl_sum implementations
can get a ecl_file handle which does not contain all the required keywords.